Remove Momentjs, Refactor Chat Input, and No Tags Quickfix
