+++ package.json
@@ -9,7 +9,6 @@
     "firebase": "^5.6.0",
     "material-ui": "^0.20.2",
     "material-ui-chip-input": "^1.0.0-beta.9",
-    "moment": "^2.22.2",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-firebase-file-uploader": "^2.4.2",
